queue_management/requirements.txt
django
django-cors-headers
django-dotenv
djangorestframework
djangorestframework-simplejwt